+++ clients/aws-sdk-transcribe-streaming/docs/conf.py
@@ -7,7 +7,7 @@ sys.path.insert(0, os.path.abspath(".."))
 
 project = "Amazon Transcribe Streaming Service"
 author = "Amazon Web Services"
-release = "0.0.1"
+release = "0.2.0"
 
 extensions = ["sphinx.ext.autodoc", "sphinx.ext.viewcode"]