+++ clients/aws-sdk-transcribe-streaming/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 ## Unreleased
 
+## v0.2.0
+
+### API Changes
+* This release adds support for additional locales in AWS transcribe streaming.
+
+### Enhancements
+* Add Standard Retry Mode
+
+### Dependencies
+* **Updated**: `smithy_aws_core[eventstream, json]` from `~=0.1.0` to `~=0.2.0`.
+* **Updated**: `smithy_core` from `~=0.1.0` to `~=0.2.0`.
+* **Updated**: `smithy_http[awscrt]~=0.3.0` from `~=0.2.0` to `~=0.3.0`.
+
 ## v0.1.0
 
 ### Features